Struct combine_language::Identifier

pub struct Identifier<PS, P> where PS: Parser<Output=char>, P: Parser<Input=PS::Input, Output=char> {
    pub start: PS,
    pub rest: P,
    pub reserved: Vec<Cow<'static, str>>,
}

Defines how to define an identifier (or operator)

Fields
start: PS

Parses a valid starting character for an identifier

rest: P

Parses the rest of the characthers in a valid identifier

reserved: Vec<Cow<'static, str>>

A number of reserved words which cannot be identifiers